Mahabir/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.1.12
django-cors-headers==3.14.0
django-datatable==0.3.1
django-enumfield==3.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.4
mongo==0.2.0
pymongo==3.12.3
pytz==2022.7
razorpay==1.3.0
requests==2.28.2
sqlparse==0.2.4
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
